Refactor br_fifo_ctrl_1r1w into two major submodules (#45)
Planning ahead for various combinations of FIFO features:
* CDC
* Push credit/valid
* Pop credit/valid
* Pipelined RAM access

Refactored br_fifo_ctrl_1r1w to move in this direction.

A general theme among the above feature variants is that we need to
decouple push and pop FSM logic. In a CDC FIFO they are on different
clock domains and there is latency to communicate between them. Decided
to start making that communication channel structurally obvious, though
right now it is overkill.

Additionally, this will make things easier when we swap out push or pop
controller for a credit-valid version.

Unfortunately, there is significant complexity needed to support key
optimizations in the common case. Right now we have redundant flop state
in a synchronous FIFO with 0 latency between push and pop controllers.