br_delay: add new output port out_stage (#79) (#92)
Adds a new output port "out_stages" which includes the output of each
stage in a delay chain. This allows the module to be used as a simple
delay element or as a shift register.

Note that out_stages[0] == in, out_stages[NumStages] == out.

For variants that include the "valid" or "valid_next" outputs, this also adds 
equivalent flags for each delay stage. 

fixes #79