Add support for non-zero read latency to FIFO controllers (#140)
Add a staging buffer to buffer up the entries from the RAM when the read
latency is positive. This staging buffer is organized as a circular
buffer with onehot-encoded read and write pointers. Some other things
added:

1. RegisterPopOutputs parameter to FIFO to allow pop_valid/pop_data to
come directly from registers at the expense of a cycle of cut-through
delay.
2. Switch br_ram_flops_1r1w_tile to br_ram_flops_1r1w and expose the
tiling and retiming parameters.
3. Add RegisterPopOutputs and latency parameter to the br_fifo_flops VCS
test suite to check the new functionality.