Fix single entry FIFO staging buffer (#248)
If EnableBypass=1, RegisterPopOutputs=1, and RamReadLatency=1, we
generate just a single register as the staging buffer.
Previously, we assumed that bypass and read data could not come
back on the same cycle, but this is incorrect. This is possible
if the buffer is not occupied. In that case, the read data should
be forwarded to the br_flow_reg_fwd and the bypass data should be
written into the buffer.