Fix data integrity bug in bypass circular buffer implementation
This fixes a bug discovered in FPV. We were failing to account for
direct push to the internal pop interface when assigning the write
enable signals.

In this case, read data returns on the same cycle as a bypass is
accepted. Both of them are targeting the same entry in the circular
buffer. However, the read data is going through directly to the
final flow reg while the bypass data is the one going to the circular
buffer entry, so there shouldn't be a conflict. The fix is two-fold:

1. Don't enable write to the circular buffer for data that is
   being directly forwarded to the internal pop interface.
2. Don't advance the write pointer for bypass beats that are
   being directly forwarded. Conversely, don't advance the read pointer
   if we are taking a bypass direct to internal pop.

commit-id:5e096379