Fix bypass cut-through latency assertion (#256)
The assertion previously assumed that bypassed data would always
go directly to the output or the final staging flop.
However, this might not be true if there are reads to the data storage
still in flight. In this case, the bypass data sits in the staging
buffer and will not go through until all the inflight read data has
gone through.