Fix initial credit return in br_credit_receiver
PR #304 introduced a bug in the configuration with RegisterPushCredit=1
that will cause one credit to be lost in the first cycle after reset
if push_credit_stall is not asserted. The decr_valid signal to the
credit counter should be gated, not just the push_credit signal to
avoid decrementing the credit counter.

commit-id:b4ade96b