amba: update axi2axil splitting logic (#416)
This PR addresses a functional bug in the AXI4 burst splitting logic for
read bursts.

In the previous implementation, we held the `arready` signal low until
we had sent all of the single-beat AXI4-Lite requests. We implicitly
made an assumption that we would be able to send all of the requests
before read responses returned. We do not buffer AXI4-Lite read
responses and pass them back through to the AXI4 initiator.

However, if the burst is really long, or if the read response data
returns quickly, then the read responses were returning to the initiator
before `arready` was even asserted. This violates the AXI4
specification.

There are a few solutions:

1. Buffer the read responses until all AXI4-Lite transactions have been
forwarded. This could lead to deadlock without using a large enough FIFO
to buffer the maximum possible burst length.
2. Assert `arready` when the first AXI4-Lite transaction is sent,
capturing the transaction attributes in a flow reg, and using the
registered transaction attributes for the remaining AXI4-Lite
transactions.

In this PR, we take Approach 2 to keep the implementation simple.