Change the behavior of br_tracker_freelist with multiple alloc ports (#429)
It would be ideal for alloc_valid to be high in strict order from LSB to
MSB when there are free entries. Previously, an entry could get "stuck"
in a staging buffer. This fixes that issue at the cost of making
alloc_valid/alloc_data unstable.

---

**Stack**:
- #441
- #437
- #430
- #429 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*