Refactor `br_mux_bin.sv` to a separate `br_mux_bin_structured_gates.sv` and mock version (#502)
`br_mux_bin.sv` has the `UseStructuredGates` parameter. When it is set
to 1, then we get trees of `br_gate_mux2` which are useful for physical
design (e.g. in CDCs) but are very slow to simulate.

We don't want to flip RTL parameters between synthesis and simulation.
And in most mux use cases people don't care about structured gates.

This change makes it like this:
* `br_mux_bin.sv`
  * Just a vanilla mux that is fine for 99% of use cases
  * No more `UseStructuredGates` parameter.
  * Module name: `br_mux_bin`
* `br_mux_bin_structured_gates.sv`
  * Uses a tree of `br_gate_mux2` for the implementation
  * Module name: `br_mux_bin_structured_gates`
* `br_mux_bin_structured_gates_mock.sv`
  * Wraps `br_mux_bin`
  * Fails to elaborate if `SYNTHESIS` preprocessor define is set.
  * Module name: `br_mux_bin_structured_gates` ❗ 

For things that require structured gates, I dropped the RTL dependency
so that downstream users can manipulate the filelist to pick up the mock
or non-mock versions as needed.

Note that we still unit test all three views of the mux library.