Fix staging buffer in shared multi-FIFO pop
The total_items input to the staging buffer is treated as
inclusive of the items in the staging buffer itself (or inflight from
the RAM). However, ram_items in the shared FIFO is actually exclusive of
the staged entries. This causes the staging buffer to incorrectly believe no
further items are available to stage, leading to underutilization of the
staging buffer. Add a parameter to deal with this properly.