Fix counter overflow in FIFO staging buffer (#600)
We were assuming that if pop_ready was true and we are attempting to
push when the staging buffer is full, we can pop on this cycle. However,
this is not the case if the staged items are still inflight from the
RAM. Need some additional logic to deal with this case when
RamReadLatency > 0.

commit-id:67a1f315