Fix incorrect integration assertion in br_credit_receiver
This assertion makes the assumption that all of the initial credit will
be held by the receiver on reset. However, initial credit could be given
to the sender, in which case this assertion would spuriously fail.

Instead of assuming how much credit the sender currently has, just track
the occupancy of the FIFO itself.

commit-id:d6456cb9