Remove combinational path from pop_ready to pop_data (#630)
In the configuration where there was no staging buffer, we tied
pop_ready to the ram_rd_addr_valid as it was being used to
increment the read address. This is undesirable as it makes the
pop_data depend on pop_ready. Break out the ram_rd_addr_valid from
the address increment so that the RAM can be continuously read as
long as data is available.