AXI (Manager) Isolator (#650)
This is the complement to the Subordinate isolator.

It keeps count of the number of expected responses from downstream, so
it can disconnect a Manager and allow it to be reset w/o messing up the
state of the downstream interconnect.

The wdata_align sub-unit was extended a bit to allow it to be used for
force-aligning AW and W (in case we disconnect the AXI Manager
mid-write).