Several fixes to axi isolators (#705)
- add simple simulation TBs that just run a clock and release reset.
This caught some issues w/ runtime/final asserts that will not be found
by FPV
- added parameter to disable wlast tracking in the resp_tracker (wlast
tracking only applicable to writes). I had been lazy and was shoving a
stream of 1s into the wlast interface for reads... which kind of worked
but then ran afoul of a bunch of assert final checks. So just did the
right thing and removed the unnecessary staging FIFOs and counters when
EnableWlastTracking=0
- there was a flow_mux select that depended directly on the RID/BID
returned from downstream, even when the RID/BID when RVALID/BVALID was
low, causing an assertion to fire because of an out-of-range value.
Forced the select to 0 when the input is not valid.
- fixed handling of unreachable FSM default cases