Add static-FIFO option for read tracker in amba_axi_isolate_sub (#732)
The dynamic fifo gets challenging for timing when the number of IDs is
large. Add an option to have dedicated FIFO per ID instead of a
dynamically shared FIFO. This is less storage efficient, but sometimes
the best compromise.

- Add support for array of static FIFOs (one per ID) instead of Dynamic
FIFO.
- Special case the AxiMaxBurstLen=1 case to remove the static FIFOs
altogether and replace with counters.
- Write tracking now *always* uses the static FIFO configuration.
Because writes are only a single response beat, no actual FIFO storage
is needed and this turns into an array of counters, rather than an array
of FIFOs. This is always more efficient than using a Dynamic FIFO.
- Remove end-to-end WLAST count. This had little purpose before (other
than to make it easy to predict exactly how many WLAST pushes could
reside within the unit to make FPV life easier) and doesn't make sense
with the static FIFO. So just getting rid of it.