Drop useless LSB of internal priority encoder of br_enc_priority_dynamic (#801)
Because of how the priority_mask works, the LSB of in_high_prio can
never be set. This causes unreachable cover points in FPV. Decrease the
width by one to avoid this.

This allows covers to be enabled on br_arb_multi_rr FPV.

---

**Stack**:
- #818
- #817
- #816
- #815
- #811
- #810
- #809
- #808
- #807
- #806
- #805
- #804
- #803
- #802
- #801 ⬅
- #800
- #799
- #798


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*