Fix multi_xfer data stability check (#829)
The instability cover didn't correctly delay the data check.
But also, writing the cover per symbol makes it very difficult to
satisfy in practice, since every position must independently be
unstable. Rewrite the assert and cover so that it checks across the
entire data.

---

**Stack**:
- #835
- #834
- #833
- #832
- #831
- #830
- #829 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*