Turn coverpoints in br_flow_checks to any-flow covers (#837)
Generally for parameter in br_flow_checks_valid_data_intg/impl, we
either assert a condition is held or cover that it is not.
However, we were doing this on a per-flow basis, which means covers
could be met for one flow but not for another. To simplify coverage
closure, turn the cover points into any-flow cover, so that any flow
satisfying the condition will meet the cover.

---

**Stack**:
- #845
- #838
- #837 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*