Remove redundant and not waivable stability assert (#857)
If EnableAssertPushValidStability is 0, and push_valid is unstable,
pop_valid will also be unstable, causing this assertion to fail. It
should really be dependent on the EnableAssertPushValidStability
parameter, however, this assertion is redundant with the valid stability
assertion within br_flow_checks_valid_data_impl, so it can just be
removed.