qualify br_ram_flops data output, add missing asserts for StructuredGates mode (#882)
added a br_gate_and2 array on data output in flops ram when used with
StructuredGates. This ensures that the output data is qualified on the
rd_data_valid before hitting any downstream combo logic (i.e. external
to BR component) before the destination flop (which might be external).

also added asserts to make sure that there's no tiling or read
pipelining in the br_ram_flops when used with StructuredGates (tile
muxing is not structured!)