Disable backpressure coverage in shared FIFO read crossbar corner case
The backpressure cover point between demux and mux is unreachable if
there happens to be only one entry mapped to the read port. Need to
disable the integration cover in the mux. The implementation cover in
the demux is a bit trickier, since it should be covered on one output
interface but not the other.

commit-id:b641c804