[csr:axil_widget] Add CSR AXI-Lite widget (#939)
This commit adds an adapter from the AXI4-Lite protocol to the Simple
CSR Bus (SCB) protocol. This is a lightweight protocol for low-bandwidth
register access allowing a single outstanding request at a time and a
configurable timeout and abort mechanism.

---

**Stack**:
- #944
- #939 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*